requirements.txt
flask==3.0.3
anthropic>=0.40.0
python-pptx==1.0.2
PyMuPDF==1.24.10
python-dotenv==1.0.1
Pillow>=10.0.0
gunicorn>=21.2.0
crewai==0.11.2
langchain-anthropic
python-docx
